backend/requirements.txt
django
django_rest_framework
django-cors-headers
psycopg2
pytest
pytest-django
